Ghost can finally chase Pacman!!

+++ Pac-Man/Pac-Man/Game.h
@@ -2,6 +2,7 @@
 #define GAME_H
 
 #include <SDL.h>
+
 #include "Renderer.h"
 #include "Pacman.h"
 #include "Ghost.h"
@@ -10,7 +11,7 @@ class Game {
     public:
         Game();
         void Run(Renderer& renderer, std::size_t target_frame_duration);
-        bool collision(short x, short y, std::array<std::array<Cell, MAP_HEIGHT>, MAP_WIDTH>& game_map, bool eat_dots);
+        bool collision(int x, int y, std::array<std::array<Cell, MAP_HEIGHT>, MAP_WIDTH>& game_map, bool eat_dots);
         int GetScore() const;
 
     private:
